Change health check port from 18080 to 18081
Because we have nginx running it's health check on 18080 which results
in a port conflict. Since we're running nginx-proxy on 8081 anyway (for
the same reason), makes sense to use 18081. Here’s the error …

```
2021/01/07 09:45:11 [emerg] 1#1: bind() to 0.0.0.0:18080 failed (98: Address in use)
nginx: [emerg] bind() to 0.0.0.0:18080 failed (98: Address in use)
2021/01/07 09:45:11 [emerg] 1#1: bind() to 0.0.0.0:18080 failed (98: Address in use)
nginx: [emerg] bind() to 0.0.0.0:18080 failed (98: Address in use)
2021/01/07 09:45:11 [emerg] 1#1: bind() to 0.0.0.0:18080 failed (98: Address in use)
nginx: [emerg] bind() to 0.0.0.0:18080 failed (98: Address in use)
2021/01/07 09:45:11 [emerg] 1#1: bind() to 0.0.0.0:18080 failed (98: Address in use)
nginx: [emerg] bind() to 0.0.0.0:18080 failed (98: Address in use)
2021/01/07 09:45:11 [emerg] 1#1: bind() to 0.0.0.0:18080 failed (98: Address in use)
nginx: [emerg] bind() to 0.0.0.0:18080 failed (98: Address in use)
2021/01/07 09:45:11 [emerg] 1#1: still could not bind()
nginx: [emerg] still could not bind()
```